+++ CHANGELOG.md
@@ -34,6 +34,7 @@ All notable changes to the Docker Language Server will be documented in this fil
   - textDocument/publishDiagnostics
     - stop flagging `BUILDKIT_SYNTAX` as an unrecognized `ARG` ([#187](https://github.com/docker/docker-language-server/issues/187))
     - use inheritance to determine if an `ARG` is truly unused ([#198](https://github.com/docker/docker-language-server/issues/198))
+    - correct range calculations for malformed variable interpolation errors ([#203](https://github.com/docker/docker-language-server/issues/203))
 
 ## [0.7.0] - 2025-05-09